requirements.txt
langchain
langgraph
langchain-openai
langchain-groq
chroma
pandas
numpy
scikit-learn
tiktoken
python-dotenv
langchain-community
arxiv
wikipedia
langchain-tavily
langgraph-cli[inmem]